backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
pyodbc==5.0.1
python-dotenv==1.0.0
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
boto3==1.34.10
azure-storage-blob==12.19.0
azure-identity==1.15.0
requests==2.31.0
werkzeug==3.0.1
